expand config and move to separate class
Extract config to separate class; at this point the motivation is mainly
to better structure the documentation for the configuration values by
pulling them out of the Client class. But this will also allow us to
move other related logic into this new class. E.g. the determination of
the default Lua system and user paths should be moved here.

* Added documentation for sharedMemory config.